migrating process to Activiti  - inscription process  - demande process  - convocation process

+++ grh/src/main/java/com/bull/grh/view/utils/MenuRHBean.java
@@ -1,15 +1,13 @@
 package com.bull.grh.view.utils;
 
-import java.io.Serializable;
-
-import javax.inject.Inject;
-
-import org.springframework.context.annotation.Scope;
-import org.springframework.stereotype.Component;
-
 import com.bull.grh.service.metier.ConvocationService;
 import com.bull.grh.service.metier.DemandeService;
 import com.bull.grh.service.metier.EntretienService;
+import org.springframework.context.annotation.Scope;
+import org.springframework.stereotype.Component;
+
+import javax.inject.Inject;
+import java.io.Serializable;
 
 @Component
 @Scope("request")
@@ -26,85 +24,85 @@ public class MenuRHBean implements Serializable {
     @Inject
     private ConvocationService convocationsService;
 
-    private long entretiensCount, startedentretiensCount, convocationsCount, startedconvocationsCount, demandesCount,
-	    starteddemandesCount;
+    private Integer entretiensCount, startedentretiensCount, convocationsCount, startedconvocationsCount, demandesCount,
+            starteddemandesCount;
 
-    public long getConvocationsCount() {
-	convocationsCount = convocationsService.getCountTaskList();
-	return convocationsCount;
+    public Integer getConvocationsCount() {
+        convocationsCount = convocationsService.getCountTaskList();
+        return convocationsCount;
     }
 
-    public void setConvocationsCount(long convocationsCount) {
-	this.convocationsCount = convocationsCount;
+    public void setConvocationsCount(Integer convocationsCount) {
+        this.convocationsCount = convocationsCount;
     }
 
-    public long getStartedconvocationsCount() {
-	startedconvocationsCount = convocationsService.getCountStartedTaskList();
-	return startedconvocationsCount;
+    public Integer getStartedconvocationsCount() {
+        startedconvocationsCount = convocationsService.getCountStartedTaskList();
+        return startedconvocationsCount;
     }
 
-    public void setStartedconvocationsCount(long startedconvocationsCount) {
-	this.startedconvocationsCount = startedconvocationsCount;
+    public void setStartedconvocationsCount(Integer startedconvocationsCount) {
+        this.startedconvocationsCount = startedconvocationsCount;
     }
 
-    public long getDemandesCount() {
-	demandesCount = demandeService.getCountDemandesSoumise();
-	return demandesCount;
+    public Integer getDemandesCount() {
+        demandesCount = demandeService.getCountDemandesSoumise();
+        return demandesCount;
     }
 
-    public void setDemandesCount(long demandesCount) {
-	this.demandesCount = demandesCount;
+    public void setDemandesCount(Integer demandesCount) {
+        this.demandesCount = demandesCount;
     }
 
-    public long getStarteddemandesCount() {
-	starteddemandesCount = demandeService.getCountStartedDemandesSoumise();
-	return starteddemandesCount;
+    public Integer getStarteddemandesCount() {
+        starteddemandesCount = demandeService.getCountStartedDemandesSoumise();
+        return starteddemandesCount;
     }
 
-    public void setStarteddemandesCount(long starteddemandesCount) {
-	this.starteddemandesCount = starteddemandesCount;
+    public void setStarteddemandesCount(Integer starteddemandesCount) {
+        this.starteddemandesCount = starteddemandesCount;
     }
 
     public EntretienService getEntretienService() {
-	return entretienService;
+        return entretienService;
     }
 
     public void setEntretienService(EntretienService entretienService) {
-	this.entretienService = entretienService;
+        this.entretienService = entretienService;
     }
 
-    public long getEntretiensCount() {
-	entretiensCount = entretienService.getCountRHTaskList();
-	return entretiensCount;
+    public Integer getEntretiensCount() {
+        entretiensCount = entretienService.getCountRHTaskList();
+        return entretiensCount;
     }
 
-    public void setEntretiensCount(long entretiensCount) {
-	this.entretiensCount = entretiensCount;
+    public void setEntretiensCount(Integer entretiensCount) {
+        this.entretiensCount = entretiensCount;
     }
 
-    public long getStartedentretiensCount() {
-	startedentretiensCount = entretienService.getCountRHStartedTaskList();
-	return startedentretiensCount;
+    public Integer getStartedentretiensCount() {
+        startedentretiensCount = entretienService.getCountRHStartedTaskList();
+        return startedentretiensCount;
     }
 
-    public void setStartedentretiensCount(long startedentretiensCount) {
-	this.startedentretiensCount = startedentretiensCount;
+    public void setStartedentretiensCount(Integer startedentretiensCount) {
+        this.startedentretiensCount = startedentretiensCount;
     }
 
     public DemandeService getDemandeService() {
-	return demandeService;
+        return demandeService;
     }
 
     public void setDemandeService(DemandeService demandeService) {
-	this.demandeService = demandeService;
+        this.demandeService = demandeService;
     }
 
     public ConvocationService getConvocationsService() {
-	return convocationsService;
+        return convocationsService;
     }
 
     public void setConvocationsService(ConvocationService convocationsService) {
-	this.convocationsService = convocationsService;
+        this.convocationsService = convocationsService;
     }
 
 }
